requirements.txt
fastapi
uvicorn
torch
transformers
python-multipart
requests
Pillow
opencv-python-headless
moviepy
scipy